Fix radian_to_degree calculating degree to radian. Fix normalize(sf::Vector2<T>) to alias unitVector

+++ HLE/main.cpp
@@ -14,7 +14,7 @@ int main()
 	sf::Vector2f vec(16, 32);
 	for (size_t i = 0; i < 100; ++i)
 	{
-		vec = hle::step_to(sf::Vector2f(i, i * 2), sf::Vector2f(64, 128));
+		vec = hle::step_to(sf::Vector2f((float)i, (float)i * 2), sf::Vector2f(64.f, 128.f));
 		std::printf("Move: %s\n",
 			hle::to_string(hle::step_to(vec, sf::Vector2f(64, 128))).c_str());
